Updated storage to mapping(bytes32 => bool) public stakeRequests Updated function interfaces Updated test cases
